+++ frontend/app/screens/workout/WorkoutApiScreen.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import { View, Text, TouchableOpacity, StyleSheet } from 'react-native';
 import { Ionicons } from '@expo/vector-icons';
 import { useNavigation } from '@react-navigation/native';
-import { WorkoutScreenNavigationProp } from '../(tabs)/WorkoutStack';
+import { WorkoutScreenNavigationProp } from '../../(tabs)/WorkoutStack';
 
 //https://ionic.io/ionicons
 // used this for icons npm install react-native-vector-icons
